Add code to avoid greylisting essential projects
Add code to the querying function that will check what admin names are listed on the project. If there are none, it will assume the project is something like `openshift-web-console` and not greylist it. It also has a code stub for hardcoding outliers to this rule. This should probably be revisited in the future, but for now, this solution will probably work. Closes #28.