This is a cry for help
I am seriously fed up with this fucking application it is cursed and I don't want to work on it any more I just want it to be done oh god fuck help me please.

Also this implements a command line interface so the docker image should be happy finally.